setup auth protected route
