+++ app/sign-in.tsx
@@ -1,14 +1,10 @@
 import { StyleSheet } from 'react-native';
+import { Text, View } from '../components/Themed';
 
-import EditScreenInfo from '../../components/EditScreenInfo';
-import { Text, View } from '../../components/Themed';
-
-export default function TabOneScreen() {
+export default function SignInScreen() {
   return (
     <View style={styles.container}>
-      <Text style={styles.title}>Tab One</Text>
-      <View style={styles.separator} lightColor="#eee" darkColor="rgba(255,255,255,0.1)" />
-      <EditScreenInfo path="app/(tabs)/index.tsx" />
+      <Text>It's a Sign In screen</Text>
     </View>
   );
 }
